[M1_TR-29] Add Link to Navbar and fix metadata error in docker
