+++ web/src/app/layout.tsx
@@ -1,18 +1,8 @@
-'use client'
-
-import { useState } from 'react';
 import type { Metadata } from 'next'
-import { Open_Sans } from 'next/font/google'
-import { Box, Container, ThemeProvider } from '@mui/material'
+import { ThemeProvider } from '@mui/material'
 
 import { theme } from '@/assets/theme'
-import Header from '@/components/organisms/Header';
-import NavBar from '@/components/organisms/Navbar';
-
-const openSansFont = Open_Sans({
-  subsets: ["latin"],
-  weight: ["300", "400", "500", "600", "700"],
-});
+import Wrapper from './wrapper';
 
 export const metadata: Metadata = {
   title: 'Create Next App',
@@ -24,30 +14,10 @@ export default function RootLayout({
 }: {
   children: React.ReactNode
 }) {
-  const [isExpanded, setIsExpanded] = useState<boolean>(true)
-
-  const handleToggle = () => {
-    setIsExpanded(!isExpanded)
-  }
-
   return (
     <html lang="en">
       <ThemeProvider theme={theme}>
-        <Box className={openSansFont.className} component={'body'} sx={{
-          margin: 0,
-          backgroundColor: 'background-primary'
-        }}>
-          <Header />
-          <Box display={'flex'}>
-            <NavBar expanded={isExpanded} handleToggle={handleToggle}/>
-            <Container maxWidth={false} disableGutters sx={{
-              marginLeft: isExpanded ? '200px' : '66px',
-              transition: 'margin 0.3s'
-            }}>
-              {children}
-            </Container>
-          </Box>
-        </Box>
+        <Wrapper>{children}</Wrapper>
       </ThemeProvider>
     </html>
   )
