+++ web/src/components/molecules/NavBarItem.tsx
@@ -1,3 +1,4 @@
+import Link from 'next/link'
 import React, { FC } from 'react'
 import { Box, Typography } from '@mui/material'
 import {SvgIconComponent} from '@mui/icons-material'
@@ -10,6 +11,7 @@ interface Props {
     height?: number | string,
     paddingX?: number | string,
     gap?: number,
+    linkTo?: string
 }
 
 const NavBarItem: FC<Props> = ({
@@ -19,40 +21,43 @@ const NavBarItem: FC<Props> = ({
     width = 200,
     height = 40,
     paddingX = 2,
-    gap = 1
+    gap = 1,
+    linkTo = "/"
     }) => {
 
     const backgroundColor = active ? 'primary.100' : 'primary.700'
     const textColor = active?  'dark': 'white'
 
     return(
-        <Box sx={{
-            display: "flex",
-            alignItems: "center",
-            width,
-            height,
-            backgroundColor,
-            ":hover": {
-                backgroundColor: 'primary.100',
-                "> div > span": {
-                    color: 'dark',
-                },
-                "> div > svg": {
-                    color: 'dark',
-                },
-            }
-        }}>
+        <Link href={linkTo} style={{textDecoration: 'none'}}>
             <Box sx={{
-            display: "flex",
-            alignItems: "center",
-            paddingX,
-            gap,
-        }}>
+                display: "flex",
+                alignItems: "center",
+                width,
+                height,
+                backgroundColor,
+                ":hover": {
+                    backgroundColor: 'primary.100',
+                    "> div > span": {
+                        color: 'dark',
+                    },
+                    "> div > svg": {
+                        color: 'dark',
+                    },
+                }
+            }}>
+                <Box sx={{
+                display: "flex",
+                alignItems: "center",
+                paddingX,
+                gap,
+            }}>
 
-            <Icon htmlColor={textColor} />
-            <Typography variant='label1r' color={textColor}>{label}</Typography>
+                <Icon htmlColor={textColor} />
+                <Typography variant='label1r' color={textColor}>{label}</Typography>
+                </Box>
             </Box>
-        </Box>
+        </Link>
     )
 }
